+++ tests/capture/test_capture_transforms.py
@@ -17,7 +17,6 @@ This submodule tests transforms with program capture
 # pylint: disable=protected-access
 from functools import partial
 
-import numpy as np
 import pytest
 
 import pennylane as qml
@@ -25,9 +24,6 @@ from pennylane.transforms.core import transform
 
 jax = pytest.importorskip("jax")
 jnp = pytest.importorskip("jax.numpy")
-
-# pylint: disable=wrong-import-position
-from pennylane.capture.transforms import MapWires
 
 pytestmark = [pytest.mark.jax, pytest.mark.usefixtures("enable_disable_plxpr")]
 
@@ -202,255 +198,3 @@ class TestCaptureTransforms:
 
         assert qfunc_jaxpr.eqns[1].primitive == qml.Z._primitive
         assert qfunc_jaxpr.eqns[2].primitive == qml.measurements.ExpectationMP._obs_primitive
-
-
-# pylint: disable=protected-access, expression-not-assigned
-class TestMapWiresTransform:
-    """Unit tests for the MapWiresTransform class."""
-
-    def test_qnode_simple(self):
-        """Test that a simple qnode is transformed correctly."""
-
-        const = 0.1
-
-        @MapWires(wire_map={0: 1, 1: 2, 2: 3, 3: 2})
-        @qml.qnode(qml.device("default.qubit", wires=4))
-        def f(x):
-            qml.RZ(x, 0)
-            qml.PhaseShift(const, 1)
-            qml.Hadamard(2)
-            return qml.expval(qml.PauliZ(3)), qml.probs(wires=[1])
-
-        jaxpr = jax.make_jaxpr(f)(0.1)
-        inner_jaxpr = jaxpr.eqns[0].params["qfunc_jaxpr"]
-
-        assert len(inner_jaxpr.eqns) == 6
-
-        assert inner_jaxpr.eqns[0].primitive == qml.RZ._primitive
-        assert inner_jaxpr.eqns[0].invars[-1].val == 1
-
-        assert inner_jaxpr.eqns[1].primitive == qml.PhaseShift._primitive
-        assert inner_jaxpr.eqns[1].invars[-1].val == 2
-
-        assert inner_jaxpr.eqns[2].primitive == qml.Hadamard._primitive
-        assert inner_jaxpr.eqns[2].invars[-1].val == 3
-
-        assert inner_jaxpr.eqns[3].primitive == qml.PauliZ._primitive
-        assert inner_jaxpr.eqns[3].invars[-1].val == 2
-
-        assert inner_jaxpr.eqns[4].primitive == qml.measurements.ExpectationMP._obs_primitive
-
-        assert inner_jaxpr.eqns[5].primitive == qml.measurements.ProbabilityMP._wires_primitive
-        assert inner_jaxpr.eqns[5].invars[-1].val == 2
-
-    def test_qnode_nested_ops(self):
-        """Test that a qnode with nested operations is transformed correctly."""
-
-        const = 0.1
-
-        @MapWires(wire_map={0: 1, 1: 2})
-        @qml.qnode(qml.device("default.qubit", wires=4))
-        def transformed_circuit(x):
-            qml.RZ(const, 0) + qml.RX(x, 0)
-            qml.PhaseShift(const, 1) @ qml.RX(x, 0) @ qml.X(0)
-            qml.Hadamard(2) @ qml.RX(x, 0) + 4 * qml.RX(const, 0)
-            return qml.state()
-
-        jaxpr = jax.make_jaxpr(transformed_circuit)(0.1)
-        inner_jaxpr = jaxpr.eqns[0].params["qfunc_jaxpr"]
-
-        assert len(inner_jaxpr.eqns) == 15
-
-        assert inner_jaxpr.eqns[0].primitive == qml.RZ._primitive
-        assert inner_jaxpr.eqns[0].invars[-1].val == 1
-
-        assert inner_jaxpr.eqns[1].primitive == qml.RX._primitive
-        assert inner_jaxpr.eqns[1].invars[-1].val == 1
-
-        assert inner_jaxpr.eqns[2].primitive == qml.ops.Sum._primitive
-
-        assert inner_jaxpr.eqns[3].primitive == qml.PhaseShift._primitive
-        assert inner_jaxpr.eqns[3].invars[-1].val == 2
-
-        assert inner_jaxpr.eqns[4].primitive == qml.RX._primitive
-        assert inner_jaxpr.eqns[4].invars[-1].val == 1
-
-        assert inner_jaxpr.eqns[5].primitive == qml.ops.Prod._primitive
-
-        assert inner_jaxpr.eqns[6].primitive == qml.X._primitive
-        assert inner_jaxpr.eqns[6].invars[-1].val == 1
-
-        assert inner_jaxpr.eqns[7].primitive == qml.ops.Prod._primitive
-
-        assert inner_jaxpr.eqns[8].primitive == qml.Hadamard._primitive
-        assert inner_jaxpr.eqns[8].invars[-1].val == 2
-
-        assert inner_jaxpr.eqns[9].primitive == qml.RX._primitive
-        assert inner_jaxpr.eqns[9].invars[-1].val == 1
-
-        assert inner_jaxpr.eqns[10].primitive == qml.ops.Prod._primitive
-
-        assert inner_jaxpr.eqns[11].primitive == qml.RX._primitive
-        assert inner_jaxpr.eqns[11].invars[-1].val == 1
-
-        assert inner_jaxpr.eqns[12].primitive == qml.ops.SProd._primitive
-        assert inner_jaxpr.eqns[13].primitive == qml.ops.Sum._primitive
-
-        assert inner_jaxpr.eqns[14].primitive == qml.measurements.StateMP._wires_primitive
-
-    def test_qnode_controlled_ops(self):
-        """Test that a qnode with controlled operations is transformed correctly."""
-
-        @MapWires(wire_map={0: 1, 1: 2, 2: 3, 3: 2})
-        @qml.qnode(qml.device("default.qubit", wires=4))
-        def f(x):
-            qml.CNOT(wires=[0, 1])
-            qml.CY(wires=[1, 2])
-            qml.CZ(wires=[2, 3])
-            qml.CRX(x, wires=[0, 1])
-            qml.CRY(x, wires=[1, 2])
-            qml.CRZ(x, wires=[2, 3])
-            return qml.var(
-                qml.QubitUnitary(
-                    np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, 1], [0, 0, 1, 0]]), wires=[0, 1]
-                )
-            )
-
-        jaxpr = jax.make_jaxpr(f)(0.1)
-        inner_jaxpr = jaxpr.eqns[0].params["qfunc_jaxpr"]
-
-        assert len(inner_jaxpr.eqns) == 8
-
-        assert inner_jaxpr.eqns[0].primitive == qml.CNOT._primitive
-        assert inner_jaxpr.eqns[0].invars[0].val == 1
-        assert inner_jaxpr.eqns[0].invars[1].val == 2
-
-        assert inner_jaxpr.eqns[1].primitive == qml.CY._primitive
-        assert inner_jaxpr.eqns[1].invars[0].val == 2
-        assert inner_jaxpr.eqns[1].invars[1].val == 3
-
-        assert inner_jaxpr.eqns[2].primitive == qml.CZ._primitive
-        assert inner_jaxpr.eqns[2].invars[0].val == 3
-        assert inner_jaxpr.eqns[2].invars[1].val == 2
-
-        assert inner_jaxpr.eqns[3].primitive == qml.CRX._primitive
-        assert inner_jaxpr.eqns[3].invars[1].val == 1
-        assert inner_jaxpr.eqns[3].invars[2].val == 2
-
-        assert inner_jaxpr.eqns[4].primitive == qml.CRY._primitive
-        assert inner_jaxpr.eqns[4].invars[1].val == 2
-        assert inner_jaxpr.eqns[4].invars[2].val == 3
-
-        assert inner_jaxpr.eqns[5].primitive == qml.CRZ._primitive
-        assert inner_jaxpr.eqns[5].invars[1].val == 3
-        assert inner_jaxpr.eqns[5].invars[2].val == 2
-
-        assert inner_jaxpr.eqns[6].primitive == qml.QubitUnitary._primitive
-        assert inner_jaxpr.eqns[6].invars[1].val == 1
-        assert inner_jaxpr.eqns[6].invars[2].val == 2
-
-        assert inner_jaxpr.eqns[7].primitive == qml.measurements.VarianceMP._obs_primitive
-
-    def test_qnode_for_loop(self):
-        """Test that a qnode with a for loop is transformed correctly."""
-
-        @MapWires(wire_map={0: 1})
-        @qml.qnode(qml.device("default.qubit", wires=4))
-        def f(x):
-            @qml.for_loop(3)
-            def g(i):
-                qml.RZ(x, 0)
-
-            g()
-            return qml.expval(qml.PauliZ(3))
-
-        jaxpr = jax.make_jaxpr(f)(0.1)
-        inner_jaxpr = jaxpr.eqns[0].params["qfunc_jaxpr"]
-        loop_jaxpr = inner_jaxpr.eqns[0].params["jaxpr_body_fn"]
-
-        assert len(loop_jaxpr.eqns) == 1
-
-        assert loop_jaxpr.eqns[0].primitive == qml.RZ._primitive
-        assert loop_jaxpr.eqns[0].invars[-1].val == 1
-
-    def test_qnode_while_loop(self):
-        """Test that a qnode with a while loop is transformed correctly."""
-
-        @MapWires(wire_map={0: 1})
-        @qml.qnode(qml.device("default.qubit", wires=4))
-        def f(x):
-            @qml.while_loop(lambda i: i < 3)
-            def g(i):
-                qml.RZ(x, 0)
-                return i + 1
-
-            g(0)
-            return qml.expval(qml.PauliZ(3))
-
-        jaxpr = jax.make_jaxpr(f)(0.1)
-        inner_jaxpr = jaxpr.eqns[0].params["qfunc_jaxpr"]
-        loop_jaxpr = inner_jaxpr.eqns[0].params["jaxpr_body_fn"]
-
-        assert len(loop_jaxpr.eqns) == 2
-
-        assert loop_jaxpr.eqns[0].primitive == qml.RZ._primitive
-        assert loop_jaxpr.eqns[0].invars[-1].val == 1
-
-    def test_qnode_conditional(self):
-        """Test that a qnode with a conditional is transformed correctly."""
-
-        @MapWires(wire_map={0: 1})
-        @qml.qnode(qml.device("default.qubit", wires=4))
-        def f(x):
-            @qml.cond(x > 0.5)
-            def g():
-                qml.RZ(x, 0)
-
-            g()
-            return qml.expval(qml.PauliZ(3))
-
-        jaxpr = jax.make_jaxpr(f)(0.1)
-        inner_jaxpr = jaxpr.eqns[0].params["qfunc_jaxpr"]
-        cond_jaxpr = inner_jaxpr.eqns[1].params["jaxpr_branches"][0]
-
-        assert len(cond_jaxpr.eqns) == 1
-
-        assert cond_jaxpr.eqns[0].primitive == qml.RZ._primitive
-        assert cond_jaxpr.eqns[0].invars[-1].val == 1
-
-    def test_invalid_wire_map_keys(self):
-        """Test that invalid wire mappings raise an error."""
-        with pytest.raises(ValueError, match="Wire map keys must be integer constants"):
-            MapWires(wire_map={"a": 1, 1: 2})
-
-    def test_invalid_wire_map_values(self):
-        """Test that invalid wire mappings raise an error."""
-        with pytest.raises(ValueError, match="Wire map values must be integer constants"):
-            MapWires(wire_map={0: "a", 1: 2})
-
-    def test_qnode_batched_parameters(self):
-        """Test qnode transformation with batched parameters."""
-
-        @MapWires(wire_map={0: 1, 1: 2})
-        @qml.qnode(qml.device("default.qubit", wires=4))
-        def f(x):
-            qml.RZ(x, 0)
-            qml.PhaseShift(0.1, 1)
-            return qml.expval(qml.PauliZ(2))
-
-        x = jax.numpy.array([0.1, 0.2, 0.3])
-        jaxpr = jax.make_jaxpr(f)(x)
-        inner_jaxpr = jaxpr.eqns[0].params["qfunc_jaxpr"]
-
-        assert len(inner_jaxpr.eqns) == 4
-
-        assert inner_jaxpr.eqns[0].primitive == qml.RZ._primitive
-        assert inner_jaxpr.eqns[0].invars[-1].val == 1
-
-        assert inner_jaxpr.eqns[1].primitive == qml.PhaseShift._primitive
-        assert inner_jaxpr.eqns[1].invars[-1].val == 2
-
-        assert inner_jaxpr.eqns[2].primitive == qml.PauliZ._primitive
-        assert inner_jaxpr.eqns[2].invars[-1].val == 2
-
-        assert inner_jaxpr.eqns[3].primitive == qml.measurements.ExpectationMP._obs_primitive
